requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
blis==0.7.4
bs4==0.0.1
cachetools==4.2.2
catalogue==2.0.4
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
cymem==2.0.5
datefinder==0.7.1
defusedxml==0.7.1
diff-match-patch==20200713
discord==1.7.3
discord.py==1.7.3
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.7
django-autocomplete-light==3.8.2
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.5.0
django-knox-rest==4.1.0
djangorestframework==3.12.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz
et-xmlfile==1.1.0
filelock==3.0.12
future==0.18.2
google-api-core==2.0.1
google-api-python-client==2.22.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.19.1
idna==2.10
Jinja2==3.0.1
json5==0.9.5
lxml==4.6.3
MarkupPy==1.14
MarkupSafe==2.0.1
mediapipe==0.8.4.2
MouseInfo==0.1.3
multidict==5.1.0
murmurhash==1.0.5
numpy==1.20.3
oauthlib==3.1.1
odfpy==1.4.1
opencv-contrib-python==4.5.2.52
opencv-python==4.5.2.52
openpyxl==3.0.7
packaging==20.9
pathy==0.5.2
Pillow==8.2.0
preshed==3.0.5
protobuf==3.17.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.7.4
pygame==2.0.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.2
pypng==0.0.20
PyQRCode==1.2.1
PyRect==0.1.4
PyScreeze==0.1.27
python-bidi==0.4.2
python-dateutil==2.8.2
PyTweening==1.0.3
pytz==2021.1
pywhatkit==4.7
PyYAML==5.4.1
regex==2021.9.24
reportlab==3.6.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
schedule==1.1.0
selenium==3.141.0
six==1.16.0
smart-open==3.0.0
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.5
sqlparse==0.4.1
srsly==2.4.1
tablib==3.0.0
textwrap3==0.9.2
thinc==8.0.3
toml==0.10.2
tqdm==4.61.0
typer==0.3.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.6
wasabi==0.8.2
webencodings==0.5.1
whitenoise==5.3.0
wikipedia==1.4.0
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
yarl==1.6.3
